refactor : cardNetwork를 구하는 함수 분리 및 CardNetwork 컴포넌트 props 형식 변경
