+++ src/components/CardDisplay/CardBody/CardBody.tsx
@@ -1,7 +1,7 @@
 import type CardExpirationDate from '@/types/CardExpirationDate';
 import { CardNumbers } from '@/types/CardNumbers';
 import CardNumberDisplay from './CardNumberDisplay';
-import styles from './cardDisplay.module.css';
+import styles from '../cardDisplay.module.css';
 
 type CardBodyProps = {
   cardNumbers: CardNumbers;
