+++ src/components/CardDisplay/CardHeader/CardHeader.tsx
@@ -1,6 +1,7 @@
 import { CardNumbers } from '@/types/CardNumbers';
-import styles from './cardDisplay.module.css';
+import styles from '../cardDisplay.module.css';
 import CardNetwork from './CardNetwork';
+import getCardNetwork from './getCardNetwork';
 
 type CardHearProps = {
   cardNumbers: CardNumbers;
@@ -10,7 +11,7 @@ const CardHeader = ({ cardNumbers }: CardHearProps) => {
   return (
     <div className={styles.cardHeader}>
       <div className={styles.chip}></div>
-      <CardNetwork cardNumbers={cardNumbers} />
+      <CardNetwork cardNetwork={getCardNetwork(cardNumbers)} />
     </div>
   );
 };
